fix: search results are now being rendered with the missing "shelf" property
The bug: A small bug was that the data fetched on search did not include
the "shelf" property which made it impossible to categorize the
books. Also some books did not have the "shelf" property on them.

The fix:
A proxy was added to update the incoming book and add the missing
"shelf" property.

A filter was added to prevent the books which did not have a
thumbnail from being loaded.