fix: Options are now highlighted in search view
Now the select options dropdown renders the categories
correctly for each book's location on the bookcase. This
means that if we search for a book that we already reading,
want to read or read, it will be highlighted so as we have an
indication of its status.

eg. If we search for a book "Pro React" we should have the option
"read" highlighted if we read that book in the past.