fix: Options are now properly displayed
An issue that existed in many of the previous commits was
that the category "currently reading" was applied to all of the
books apart from the fact that some of those books were in an
other category.

The inproper handle of the status of the currentl selected book
made the books render correctly in the shelves but with the wrong
status. That meant that the status of the books was undefined
during the rendering process of the shelves.